multimodel_count_regression()

Multimodel Count Regression

multimodel_count_regression(
  outcomes_interest,
  models,
  plot_density,
  plot_qq,
  estimate,
  avg_effect,
  rootogram,
  full_model,
  null_model,
  data,
  workers,
  path
)

Arguments

outcomes_interest

A character vector specifying the outcomes of interest.

models

A character vector specifying the models to be fitted.

plot_density

A logical value indicating whether to plot density plots of the fitted models.

plot_qq

A logical value indicating whether to plot Q-Q plots of the fitted models.

estimate

A logical value indicating whether to estimate model parameters and standard errors.

avg_effect

A logical value indicating whether to compute average marginal effects.

rootogram

A logical value indicating whether to generate rootograms for the fitted models.

full_model

The formula for the full model.

null_model

The formula for the null model.

data

The data frame containing the variables of interest.

workers

The number of parallel workers to use for parallel computing.

path

The path to save the plots.

Value

A list containing the results of the multimodel count regression analysis.

Details

This function performs multimodel count regression analysis.