Selecting tuning parameters for sparse CCA

Selecting tuning parameters for sparse CCA

parallel_cca_permute(
  x,
  z,
  obj,
  penaltyxs = NULL,
  penaltyzs = NULL,
  niter = 3,
  v = NULL,
  trace = FALSE,
  nperms = 25,
  standardize = TRUE,
  upos = FALSE,
  uneg = FALSE,
  vpos = FALSE,
  vneg = FALSE,
  outcome = NULL,
  y = NULL,
  cens = NULL
)

Arguments

x

Expression-driven adjacency matrix

z

Spatial-driven adjacency matrix

obj

stCCA S3 Object

penaltyxs

-desc

penaltyzs

-desc

niter

The number of iterations to be performed. Default - 20.

v

-desc

trace

-desc

nperms

-desc

standardize

-desc

upos

-desc

uneg

-desc

vpos

-desc

vneg

-desc

outcome

-desc

y

-desc

cens

-desc

References

We adapted and modified the code from the CCA function in the PMA package to be compatible and scalable with single-cell multimodal integration through parallel computing.